+++ upload-service/src/app.module.ts
@@ -2,10 +2,18 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { UploadsModule } from './uploads/uploads.module';
+import { StatusService } from './status/status.service';
+import { ConfigModule } from '@nestjs/config';
+import { StatusController } from './status/status.controller';
 
 @Module({
-  imports: [UploadsModule],
-  controllers: [AppController],
-  providers: [AppService],
+  imports: [
+    ConfigModule.forRoot({
+      isGlobal: true,
+    }),
+    UploadsModule,
+  ],
+  controllers: [AppController, StatusController],
+  providers: [AppService, StatusService],
 })
 export class AppModule {}